refactor: move generic db operations into db-actions
